requirements.txt
# Audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
webrtcvad>=2.0.10

# Speech-to-Text
faster-whisper>=0.10.0

# LLM Integration
ollama>=0.1.0
requests>=2.31.0

# Text-to-Speech (Piper will be installed separately)
# Piper TTS - binary installation via script

# System integration
psutil>=5.9.0
beautifulsoup4>=4.12.0
python-xdg>=0.28

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic-settings>=2.1.0

# Terminal UI
rich>=13.7.0

# Error Recovery (Phase 1)
tenacity>=8.2.0
openai-whisper>=20230918  # Fallback STT
SpeechRecognition>=3.10.0  # Alternative STT

# Wake Word Detection (Phase 1)
pvporcupine>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0
freezegun>=1.2.0  # Time mocking for tests
